+++ port/port-game-server-tcp/src/main/java/club/ss220/port/gameserver/tcp/paradise/presentation/ParadiseOnlineStaffStatusPresentation.java
@@ -46,8 +46,8 @@ public record ParadiseOnlineStaffStatusPresentation(
     private static Duration parseAfk(Map<String, Object> data) {
         return Optional.ofNullable(data.get(AFK_PROPERTY))
                 .map(String::valueOf)
-                .map(Integer::parseInt)
-                .map(deciseconds -> Duration.ofMillis(deciseconds * 100L))
+                .map(Double::parseDouble)
+                .map(deciseconds -> Duration.ofMillis(deciseconds.longValue() * 100))
                 .orElseThrow(() -> propertyNotFound(AFK_PROPERTY));
     }
